add dependencies for running the postgres tap infrastructure (#51)
To be able to run postgres' tap test framework we need to have the `IPC::Run` library for perl installed in our test container. Since this package is not directly installable from `apt` we install `cpanminus` that comes with/is an installer for perl modules.

Subsequently we install the required perl library.