Fix the installation of postgres when more then 1 release is available
For postgres 13.4 there are 2 concurrent released versions:
 - 13.4-1
 - 13.4-4

The bash pipeline to infer the version to use had troubles parsing the versions
with more than 1 package per postgres minor version.

This PR changes the bash pipeline to correctly choose 1. For now we just take
the first one it finds with the use of `head -n1`.