Bump mitmproxy version (#65)
What started as a simple change to upgrade the mitmproxy version turned out to be more involved.

We were still running on python 3.6 which didn't have the patched version of mitmproxy released.
To upgrade python to 3.9 we needed to also update debian from stretch to bullseye.
Which in turn caused zstd and lz4 to be upgraded causing test failures due to difference in compressed sizes.

Long story short, all dependencies have been upgraded!

For mitmproxy we couldn't use the released version 7.0.4 as it has a bug that prevents `TCPFlow.kill` to not work. Given our heavy reliance on that API for our test suite we have, as advised by the maintainers, patched mitmproxy in a fork to have a working implementation of `TCPFlow.kill` until this gets addressed in mitmproxy.

Lastly, given the fierce changes to the ci images, we decided it became time to version the images. We settled on using a `-vYYYY_mm_dd` suffix to the image names. Versioning the image to the date we build the images. To make dev stand out a bit easier from it I decided on small changes to the dev tags for easier readability.