Parallelize builds (#100)
Previously the dev build pipelines would build the images for each PG
version sequentially. This changes it to build them in parallel. To make
sure that we only have to update postgres versions in one place, these
are now moved to a separate file, which is then read by both the CI and
the makefile.