Install dev requirements in stylechecker (#109)
Needed to check Python formatting and linting from CI for
https://github.com/citusdata/citus/pull/6700

Since some of these tools need a somewhat recent python version this
also required an update of the alpine base image. And to make that in
turn work, we need to install `uncrustify` from source (just like we
recommend devs to do).