Move CI images to GHCR (#157)
The main motivation of this patch is to move the images to the github
container registry. But as quite often the case, the images were not
building, thus some changes needed to be made to get the images to
build.

- Python construct package is bumped. There was a dependency failing to
install, which succeeds on a newer version. This change is made by
changing the version on the citus repo, and re-export the
requirements.txt
- pin the `postgresql-$PG_MAJOR` version to the same pgdg version. For
pg16.2 new packages have been created. This happens once in a while and
to counter that we already pin most of the postgres ecosystem packages
to a specific pgdg version. Somehow `postgresql-...` wasn't part of the
pinning, causing some errors during image builds.

The scripts used the old syntax of multi line run blocks, requiring a
lot of escaping and chaining of lines with the trailing `\` character.
Newer versions of docker allow an easier to maintain and readable syntax
via
```Dockerfile
RUN <<'EOF'
# first comment will be shown during build as 'name' of the task
...
EOF
```

Given I already needed to change the body of many of the RUN blocks I
took the liberty to refactor this at the same time.

---------

Co-authored-by: Hanefi Onaldi <Hanefi.Onaldi@microsoft.com>